mod_handler: fix critical bugs in LoRa and power management
- Fix power_mv type uint8_t -> uint32_t (battery voltage truncation)
- Add buffer overflow check in lora_data_send (max len 116)
- Fix lora_exit_at double \r\n (lora_send_at already appends it)
- Fix cmd_at AT+Z branch missing lora_mode_mutex release
- Replace lora_power_enable(false) with lora_deinit() in system_sleep
  to properly stop DMA before power off
- Restructure lora_rssi_thread mutex unlock: release before sleep
  period on normal path, use goto SLEEP_EXIT for sleep interrupt
- Set ok=false when telemetry interrupted by sleep

+++ applications/mod_handler/include/common.h
@@ -34,7 +34,7 @@ typedef struct {
 	int x_degree;
 	int y_degree;
 	uint8_t h_button;
-	uint8_t power_mv;
+	uint32_t power_mv;
 #define CAN_TYPE  1
 #define LORA_TYPE 2
 	uint8_t connect_type;
